shiki server package problem.
log drop:
 ./node_modules/.pnpm/streamdown@1.0.12_@types+react@19.1.5_react@19.1.0/node_modules/streamdown/dist
Package shiki can't be external
The request shiki matches serverExternalPackages (or the default list).
The request could not be resolved by Node.js from the project directory.
Packages that should be external need to be installed in the project directory, so they can be resolved from the output files.
Try to install it into the project directory by running npm install shiki from the project directory.